requirements.txt
streamlit
pandas
numpy
scikit.learn
matplotlib
seaborn
joblibpip